requirements.txt
anyio==4.9.0
bcrypt==4.3.0
click==8.1.8
colorama==0.4.6
fastapi==0.95.2
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.0.33
passlib==1.7.4
protobuf==3.20.3
pydantic==1.10.21
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.13.1
uvicorn==0.22.0
